define index query in one place
