+++ __test__/server/db/export.js
@@ -1,7 +1,7 @@
 require('dotenv').config({ path: '../../../../.env' })
 const config = require('../../../../knexfile.js')
 const knex = require('knex')(config)
-import { tables } from './tables'
+import { tables, indexQuery } from './utils'
 import fs from 'fs'
 
 function getSchema(s) {
@@ -11,20 +11,7 @@ function getSchema(s) {
   })
 }
 function getIndexes() {
-  // // return knex('pg_indexes')
-  // .select()
-  // .where({ schemaname: 'public' })
-  // .whereNot({ tablename: 'migrations' })
-  return knex.raw(`SELECT conrelid::regclass AS table_from
-     , conname
-     , pg_get_constraintdef(c.oid)
-FROM   pg_constraint c
-JOIN   pg_namespace n ON n.oid = c.connamespace
-WHERE  contype IN ('f', 'p ')
-AND    conrelid::regclass::text <> 'migrations'
-AND    n.nspname = 'public'
-ORDER  BY conrelid::regclass::text, contype DESC;`)
-  // this table is deprecated now
+  return knex.raw(indexQuery)
   .then(indexes => {
     fs.writeFileSync('indexes.json', JSON.stringify(indexes, null, 2))
     console.log('exported indices')
